% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_all_missing_by_group.R
\name{count_all_missing_by_group}
\alias{count_all_missing_by_group}
\title{Count Missing Values by Group}
\usage{
count_all_missing_by_group(data, group_col, .groups = "drop")
}
\arguments{
\item{data}{Data in data.frame or tibble format.}

\item{group_col}{Column in data to group by, specified without quotes.}

\item{.groups}{(OPTIONAL) Controls whether the output data frame is grouped.
By default, drops all levels of grouping ("drop"). Specify "keep" to keep
the output data frame grouped by group. See summarize() documentation in
dplyr for a full list of options.}
}
\value{
A data frame or tibble with the levels of group, and the number of NAs within
each level of group for all columns in data except group.
}
\description{
Given a data frame data and a
column group, count_all_missing_by_group() creates
a new data frame with one row per level of group.
The first column of the output data frame
contains the levels of group, and the rest of
the columns contain the number of missing values for
all columns in data except group.
}
\details{
This function is a wrapper around the group_by() and
summarize() functions in the dplyr package.
}
\examples{
# Example usage with airquality dataset
count_all_missing_by_group(airquality, Month)

# Using pipe operator
airquality |> count_all_missing_by_group(Month)

# Controlling the grouping behavior
count_all_missing_by_group(airquality, Month, .groups = "keep")

}
